rename general name crate to specific platform timer name
